requirements.txt
# Function dependencies, for example:
# package>=version
alpaca-trade-api==0.52.0
python-dotenv==0.15.0
robin-stocks==2.0.0
webull @ git+https://github.com/tedchou12/webull.git@8b57cd5fda321ed05564dc92bcb61c157049f527
